feat: pss-cln-set-legacy-packages-001 - Fold identityinventory into document
Move identity inventory implementation under document internal ownership, retarget fixture paths, and keep a transitional public shim for compile continuity.
