+++ cmd/packagetargetmanifestcheck/owners.go
@@ -300,6 +300,7 @@ var nestedOwnerMoveRules = map[string][]nestedPathRule{
 		{exact: "testproviders", prefix: "testproviders/", dest: "operator_settings/internal"},
 		{exact: "testdata", prefix: "testdata/", dest: "operator_settings/internal"},
 		{exact: "internal/service", prefix: "internal/service/", dest: "operator_settings/internal"},
+		{exact: "internal/construct", prefix: "internal/construct/", dest: "operator_settings/internal"},
 	},
 	"workers": {
 		{exact: "construction", prefix: "construction/", dest: "workers/internal/services/runtime_assembly"},
