+++ docs/internal/packaged-service-structure/package-target-manifest.json
@@ -334,8 +334,10 @@
     "pkg/services/models/wire",
     "pkg/services/operator_settings",
     "pkg/services/operator_settings/identityinventory",
+    "pkg/services/operator_settings/internal/construct",
     "pkg/services/operator_settings/internal/service",
     "pkg/services/operator_settings/internal/services/document",
+    "pkg/services/operator_settings/internal/services/document/identityinventory",
     "pkg/services/operator_settings/internal/services/document/internal/service",
     "pkg/services/operator_settings/internal/services/document/wire",
     "pkg/services/operator_settings/internal/services/resolution",
@@ -2025,12 +2027,22 @@
       "destination": "operator_settings/internal/services/document"
     },
     {
+      "packagePath": "pkg/services/operator_settings/internal/construct",
+      "disposition": "move",
+      "destination": "operator_settings/internal"
+    },
+    {
       "packagePath": "pkg/services/operator_settings/internal/service",
       "disposition": "move",
       "destination": "operator_settings/internal"
     },
     {
       "packagePath": "pkg/services/operator_settings/internal/services/document",
+      "disposition": "retain",
+      "destination": "operator_settings/internal/services/document"
+    },
+    {
+      "packagePath": "pkg/services/operator_settings/internal/services/document/identityinventory",
       "disposition": "retain",
       "destination": "operator_settings/internal/services/document"
     },
