+++ pkg/services/operator_settings/internal/services/document/identityinventory/input_index_cases.go
