+++ pkg/services/operator_settings/internal/services/document/identityinventory/testdata/baseline/system-config-input-index.json
