+++ pkg/services/operator_settings/internal/services/document/identityinventory/testdata/fixtures/invalid/malformed-json.json
