+++ pkg/services/operator_settings/internal/services/document/identityinventory/testdata/fixtures/valid/defaults-sibling.json
