+++ pkg/services/operator_settings/internal/services/document/identityinventory/testdata/fixtures/valid/empty-scope-field.json
