+++ pkg/services/operator_settings/internal/services/document/identityinventory/testdata/fixtures/valid/missing-scope.json
