+++ pkg/services/operator_settings/internal/services/document/identityinventory/testdata/fixtures/valid/whitespace-only.json
